# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "version": "1.00"
    },
    "paths": {
...
        "/api/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get current user",
                "operationId": "current_user",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/BaseResponse"
                                        },
                                        {
                                            "properties": {
                                                "result": {
                                                    "properties": {
                                                        "user": {
                                                            "$ref": "#/components/schemas/UserResource"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/user/car": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get user current car",
                "operationId": "current_car",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/BaseResponse"
                                        },
                                        {
                                            "properties": {
                                                "result": {
                                                    "properties": {
                                                        "car": {
                                                            "$ref": "#/components/schemas/CarResource"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get users list",
                "operationId": "users",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Page items limit"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/BaseResponse"
                                        },
                                        {
                                            "$ref": "#/components/schemas/PaginateResponse"
                                        },
                                        {
                                            "properties": {
                                                "result": {
                                                    "properties": {
                                                        "users": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/components/schemas/UserResource"
                                                            }
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/users/{user_uuid}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get user by uuid",
                "operationId": "user_by_uuid",
                "parameters": [
                    {
                        "name": "user_uuid",
                        "in": "path",
                        "description": "User uuid",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/BaseResponse"
                                        },
                                        {
                                            "properties": {
                                                "result": {
                                                    "properties": {
                                                        "user": {
                                                            "$ref": "#/components/schemas/UserResource"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "CarResource": {
                "title": "Car resource",
                "properties": {
                    "uuid": {
                        "schema": "CarResource",
                        "title": "UUID",
                        "description": "Car UUID",
                        "type": "string",
                        "example": "d1b0c6be-e64d-4baf-acec-17e7c6162216"
                    },
                    "company": {
                        "schema": "CarResource",
                        "title": "Company name",
                        "description": "Car company name",
                        "type": "string",
                        "example": "BMW"
                    },
                    "model_family": {
                        "schema": "CarResource",
                        "title": "Model family",
                        "description": "Car model family",
                        "type": "string",
                        "example": "X"
                    },
                    "model_number": {
                        "schema": "CarResource",
                        "title": "Model number",
                        "description": "Car model number",
                        "type": "string",
                        "example": "100"
                    },
                    "state": {
                        "schema": "CarResource",
                        "title": "State",
                        "description": "Car state",
                        "type": "string",
                        "enum": [
                            "free",
                            "busy"
                        ],
                        "example": "free"
                    }
                },
                "type": "object"
            },
            "BaseResponse": {
                "description": "base response",
                "properties": {
                    "result": {
                        "title": "Result object",
                        "type": "object"
                    },
                    "message": {
                        "title": "Error message",
                        "type": "string"
                    },
                    "errors": {
                        "title": "Errors array",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "type": "object"
            },
            "PaginateResponse": {
                "description": "paginate response",
                "properties": {
                    "result": {
                        "title": "Result object",
                        "properties": {
                            "total": {
                                "type": "integer",
                                "example": "10"
                            },
                            "limit": {
                                "type": "integer",
                                "example": "10"
                            },
                            "page": {
                                "type": "integer",
                                "example": "1"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "UserResource": {
                "title": "User resource",
                "properties": {
                    "uuid": {
                        "title": "UUID",
                        "description": "User UUID",
                        "type": "string",
                        "example": "d1b0c6be-e64d-4baf-acec-17e7c6162216"
                    },
                    "name": {
                        "title": "Name",
                        "description": "User name",
                        "type": "string",
                        "example": "Ilya"
                    },
                    "email": {
                        "title": "Email",
                        "description": "User email",
                        "type": "string",
                        "example": "test@example.com"
                    }
                },
                "type": "object"
            }
        },
...
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "access_token for access to /user and /cars",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}